feat: Move primary transport checked update into "remove transports" transaction
This guarantees that the primary transport is never removed while being in use. At least this way
it's obvious that there are no such corner cases.
